requirements.txt
Flask>=2.0
scikit-learn>=1.0
pandas>=1.3
mysql-connector-python>=8.0
joblib
numpy
